requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
pyyaml==6.0.2
typer==0.16.1
pytest==8.4.1
httpx==0.28.1
